Refactored version of the existing landscape brush of landmass
My initial painpoint was that the usual auto-paint approach for
landscapes was not allowing me to use the grass tool because no
real paint-layer weights were available. At least this
was my understanding.

In order to create these weights I tried to use the existing landmass
plugins brushes but they were somewhat hard to understand and modify.
I extracted the weight-adjustment features and, as it was similar, also
the material based height-adjustment features.

There are some masks provided which most base on a somewhat working
algorithm to compute the normal of the landscape.

I found that the brushes tend to "forget" values of the private
variables (I dont mean the local ones). This made the refactoring
really hard as this wasnt something i expected. I went ahead and
reused the original manager approach but used it only as a cache.